src/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "fastapi[standard]>=0.115.8",
    "pyjwt>=2.10.1",
    "uvicorn>=0.34.0",
]
